% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/F.norm2.R
\name{F.norm2}
\alias{F.norm2}
\title{The Squared Frobenius Norm}
\usage{
F.norm2(matrix)
}
\arguments{
\item{matrix}{a matrix}
}
\value{
a scalar of the squared Frobenius norm
}
\description{
Calculate the squared Frobenius norm of a matrix
}

