% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FinCovRegularization.R
\docType{package}
\name{FinCovRegularization}
\alias{FinCovRegularization}
\alias{FinCovRegularization-package}
\title{FinCovRegularization: Covariance Matrix Estimation and Regularization
  for Finance}
\description{
Estimation and regularization for covariance matrix of asset returns.
For covariance matrix estimation, three major types of factor models are
included: macroeconomic factor model, fundamental factor model and
statistical factor model. For covariance matrix regularization,
four regularized estimators are included: banding, tapering,
hard-thresholding and soft-thresholding. The tuning parameters of these
regularized estimators are selected via cross-validation.
}

