% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Ind.Cov.R
\name{Ind.Cov}
\alias{Ind.Cov}
\title{Independence opreator on Covariance Matrix}
\usage{
Ind.Cov(sigma)
}
\arguments{
\item{sigma}{a covariance matrix}
}
\value{
a regularized covariance matrix after applying independence model
}
\description{
Apply independence model on a covariance matrix.
}

