% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FinCovRegularization.R
\docType{data}
\name{m.excess.c10sp9003}
\alias{m.excess.c10sp9003}
\title{10 stock and S&P 500 excess returns}
\format{A matrix with 168 rows and 11 variables}
\usage{
data(m.excess.c10sp9003)
}
\description{
A dataset containing monthly excess returns of 10 stocks and
S$P 500 index return from January 1990 to December 2003
}
\keyword{datasets}

