% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Generic.CovCv.R
\name{plot.CovCv}
\alias{plot.CovCv}
\title{plot CovCv object}
\usage{
\method{plot}{CovCv}(x, ...)
}
\arguments{
\item{x}{CovCv object to plot}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Visualizes the results of covariance matrix regularization parameter tuning
}

