% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.CovCv.R
\name{print.CovCv}
\alias{print.CovCv}
\title{print CovCv object}
\usage{
\method{print}{CovCv}(x, ...)
}
\arguments{
\item{x}{CovCv object to plot}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
print selected optimal parameter
}
\keyword{internal}

