% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskParity.R
\name{RiskParity}
\alias{RiskParity}
\title{Risk Parity Portfolio}
\usage{
RiskParity(cov.mat)
}
\arguments{
\item{cov.mat}{an estimated p*p covariance matrix}
}
\value{
a numerical vector containing the estimated portfolio weights
}
\description{
Computing a Risk Parity portfolio weights from the estimated
covariance matrix of return series.
}
\examples{
data(m.excess.c10sp9003)
assets <- m.excess.c10sp9003[,1:10]
RiskParity(cov(assets))
}

