% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.CovCv.R
\name{summary.CovCv}
\alias{summary.CovCv}
\title{Display a useful description of a CovCv object}
\usage{
\method{summary}{CovCv}(object, ...)
}
\arguments{
\item{object}{CovCv object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Display a useful description of a CovCv object
}
\keyword{internal}

