% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/threshold.min.R
\name{threshold.min}
\alias{threshold.min}
\title{Minimum threshold constant}
\usage{
threshold.min(sigma, method = "hard")
}
\arguments{
\item{sigma}{a covariance matrix}

\item{method}{a character, indicating thresholding method: "soft" or "hard"}
}
\value{
minimum constant for thresholding
}
\description{
This function is for determining the minimum constant in the threshold
that guarantees the positive definiteness of the estimator.
}
\keyword{internal}

