% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FinCovRegularization.R
\docType{package}
\name{FinCovRegularization}
\alias{FinCovRegularization}
\alias{FinCovRegularization-package}
\title{FinCovRegularization: Covariance Matrix Estimation and Regularization
  for Finance}
\description{
This is a package providing three categories of important functions for
statistical analysis of financial data: including portfolio optimization:
global mimnum variance portfolio, risk parity portfolio;
covariance estimation: macroeconomic, fundamental, and
statistical factor model; covariance regularization: banding,
tapering, thresholding and corresponding cross-validation
for parameter selection
}

