% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ind.Cov.R
\name{Ind.Cov}
\alias{Ind.Cov}
\title{Independence opreator on Covariance Matrix}
\usage{
Ind.Cov(sigma)
}
\arguments{
\item{sigma}{a covariance matrix}
}
\value{
a regularized covariance matrix after applying independence model
}
\description{
Apply independence model on a covariance matrix.
}
\examples{
data(m.excess.c10sp9003)
cov.SAM <- cov(m.excess.c10sp9003)
Ind.Cov(cov.SAM)
}

