% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O.norm2.R
\name{O.norm2}
\alias{O.norm2}
\title{The Squared Operator Norm}
\usage{
O.norm2(matrix)
}
\arguments{
\item{matrix}{a matrix}
}
\value{
a scalar of the squared Operator norm
}
\description{
Calculate the squared Operator norm of a matrix
}
\examples{
data(m.excess.c10sp9003)
cov.SAM <- cov(m.excess.c10sp9003)
O.norm2(cov.SAM)
}

